+++ tests/test_widgets.py
@@ -460,7 +460,8 @@ class TestWidgetManagerInActivity(TestBetamax):
 
         widget1 = self.wm.add_signature_widget(attachment_property=picture, title="Yes, my precious",
                                                custom_undo_button_text="Remove za widget",
-                                               custom_button_text="Sign za widget")
+                                               custom_button_text="Sign za widget",
+                                               show_name_and_date=False)
         widget2 = self.wm.add_signature_widget(attachment_property=picture, title=False,
                                                custom_undo_button_text=False,
                                                custom_button_text=False)